pyproject.toml
[project]
name = "KarmaNodes"
description = "Advanced KSampler node for ComfyUI with multi-cycle sampling and progressive upscaling, as well as professional post-processing tools"
version = "1.0.0"
license = {file = "LICENSE"}
authors = [
    {name = "Karma Swint", email = "karma@karmaviz.biz"}
]
keywords = ["comfyui", "stable-diffusion", "sampling", "upscaling", "ai", "image-generation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.9"
dependencies = [
    "torch>=2.0.0",
    "Pillow>=9.0.0",
    "numpy>=1.21.0",
    "scikit-image>=0.19.0",
    "scipy>=1.9.0",  # For enhanced film grain quality
]